Move "Bulk update" func. to separate viewModel
