+++ src/ggrc-client/js/components/assessment/assessments-bulk-complete/assessments-bulk-complete.js
@@ -12,15 +12,12 @@ import '../../required-info-modal/required-info-modal';
 import canComponent from 'can-component';
 import canStache from 'can-stache';
 import canBatch from 'can-event/batch/batch';
-import canMap from 'can-map';
 import template from './assessments-bulk-complete.stache';
-import ObjectOperationsBaseVM from '../../view-models/object-operations-base-vm';
 import {STATES_KEYS} from '../../../plugins/utils/state-utils';
 import loFindIndex from 'lodash/findIndex';
 import {request} from '../../../plugins/utils/request-utils';
 import {backendGdriveClient} from '../../../plugins/ggrc-gapi-client';
 import {ggrcPost} from '../../../plugins/ajax-extensions';
-import {trackStatus} from '../../../plugins/utils/background-task-utils';
 import {confirm} from '../../../plugins/utils/modals';
 import {
   getFetchErrorInfo,
@@ -35,11 +32,7 @@ import {
 import loSome from 'lodash/some';
 import loFind from 'lodash/find';
 import {getPlainText} from '../../../plugins/ggrc-utils';
-import {
-  create,
-  setDefaultStatusConfig,
-} from '../../../plugins/utils/advanced-search-utils';
-import {getAvailableAttributes} from '../../../plugins/utils/tree-view-utils';
+import AssessmentsBulkUpdatable from '../view-models/assessments-bulk-updatable-vm';
 
 /**
  * Map of types from FE to BE format
@@ -54,7 +47,7 @@ const attributesType = {
   dropdown: 'Dropdown',
 };
 
-const viewModel = ObjectOperationsBaseVM.extend({
+const viewModel = AssessmentsBulkUpdatable.extend({
   define: {
     isSelectButtonDisabled: {
       get() {
@@ -84,17 +77,8 @@ const viewModel = ObjectOperationsBaseVM.extend({
       },
     },
   },
-  element: null,
-  showSearch: false,
   showFields: false,
-  isMyAssessmentsView: false,
-  mappedToItems: [],
-  filterItems: [],
-  defaultFilterItems: [],
-  mappingItems: [],
-  filterAttributes: [],
   statesCollectionKey: STATES_KEYS.BULK_COMPLETE,
-  type: 'Assessment',
   isAttributesGenerating: false,
   attributeFields: [],
   isAttributesGenerated: false,
@@ -109,7 +93,6 @@ const viewModel = ObjectOperationsBaseVM.extend({
       files: [],
     },
   },
-  timeoutId: null,
   isCompleting: false,
   /**
    * Contains selected objects (which have id and type properties)
@@ -242,9 +225,6 @@ const viewModel = ObjectOperationsBaseVM.extend({
     return request('/api/bulk_operations/cavs/search', {
       ids: this.getSelectedAssessmentsIds(),
     });
-  },
-  getSelectedAssessmentsIds() {
-    return this.attr('selected').serialize().map((selected) => selected.id);
   },
   updateAttributeField({field, value}) {
     field.attr('value', value);
@@ -382,31 +362,6 @@ const viewModel = ObjectOperationsBaseVM.extend({
     requiredInfoModal.attr('state.open', true);
     canBatch.stop();
   },
-  initDefaultFilter() {
-    const stateConfig = setDefaultStatusConfig(
-      new canMap,
-      this.attr('type'),
-      this.attr('statesCollectionKey')
-    );
-    const items = [
-      create.state(stateConfig),
-      create.operator('AND'),
-      create.attribute({
-        field: 'Assignees',
-        operator: '~',
-        value: GGRC.current_user.email,
-      }),
-    ];
-
-    this.attr('filterItems', items);
-    this.attr('defaultFilterItems', items);
-  },
-  initFilterAttributes() {
-    const attributes = getAvailableAttributes(this.attr('type'))
-      .filter(({attr_name: attrName}) => attrName !== 'status');
-
-    this.attr('filterAttributes', attributes);
-  },
   getValForCompleteRequest(type, value) {
     switch (type) {
       case 'checkbox':
@@ -467,17 +422,6 @@ const viewModel = ObjectOperationsBaseVM.extend({
       assessments_ids: this.getSelectedAssessmentsIds(),
     };
   },
-  trackBackgroundTask(taskId) {
-    notifier('progress', 'Your bulk update is submitted. ' +
-        'Once it is done you will get a notification. ' +
-        'You can continue working with the app.');
-    const url = `/api/background_tasks/${taskId}`;
-    const timeoutId = trackStatus(
-      url,
-      () => this.onSuccessHandler(),
-      () => this.onFailureHandler());
-    this.attr('timeoutId', timeoutId);
-  },
   completeAssessments() {
     this.attr('isCompleting', true);
     backendGdriveClient.withAuth(
@@ -521,17 +465,6 @@ const viewModel = ObjectOperationsBaseVM.extend({
     } else {
       this.completeAssessments();
     }
-  },
-  onSuccessHandler() {
-    const reloadLink = window.location.href;
-    notifier('success', 'Bulk update is finished successfully. {reload_link}',
-      {reloadLink});
-  },
-  onFailureHandler() {
-    notifier('error', 'Bulk update is failed.');
-  },
-  closeModal() {
-    this.attr('element').find('.modal-dismiss').trigger('click');
   },
   init() {
     this.initDefaultFilter();
